refactor: rename buffer_resource.h to buffer.h and update all references
