+++ orteaf/include/orteaf/internal/execution/allocator/policies/chunk_locator/direct_chunk_locator.h
@@ -8,7 +8,7 @@
 #include "orteaf/internal/base/heap_vector.h"
 #include "orteaf/internal/diagnostics/error/error.h"
 #include "orteaf/internal/diagnostics/error/error_macros.h"
-#include "orteaf/internal/execution/allocator/buffer_resource.h"
+#include "orteaf/internal/execution/allocator/buffer.h"
 #include "orteaf/internal/execution/allocator/policies/policy_config.h"
 
 namespace orteaf::internal::execution::allocator::policies {
