+++ orteaf/include/orteaf/internal/execution/allocator/policies/freelist/device_linked_freelist_policy.h
@@ -5,7 +5,7 @@
 
 #include <orteaf/internal/base/heap_vector.h>
 #include <orteaf/internal/diagnostics/error/error_macros.h>
-#include <orteaf/internal/execution/allocator/buffer_resource.h>
+#include <orteaf/internal/execution/allocator/buffer.h>
 #include <orteaf/internal/execution/allocator/policies/policy_config.h>
 #include <orteaf/internal/execution/base/execution_traits.h>
 #include <orteaf/internal/execution/execution.h>
