+++ orteaf/include/orteaf/internal/execution/allocator/pool/segregate_pool.h
@@ -2,7 +2,7 @@
 
 #include <algorithm>
 #include <limits>
-#include <orteaf/internal/execution/allocator/buffer_resource.h>
+#include <orteaf/internal/execution/allocator/buffer.h>
 #include <orteaf/internal/execution/allocator/pool/segregate_pool_stats.h>
 #include <orteaf/internal/execution/allocator/size_class_utils.h>
 #include <orteaf/internal/execution/execution.h>
