+++ orteaf/include/orteaf/internal/execution/mps/manager/mps_buffer_manager.h
@@ -13,7 +13,7 @@
 #include "orteaf/internal/base/manager/pool_manager.h"
 #include "orteaf/internal/base/pool/slot_pool.h"
 #include "orteaf/internal/diagnostics/error/error.h"
-#include "orteaf/internal/execution/allocator/buffer_resource.h"
+#include "orteaf/internal/execution/allocator/buffer.h"
 #include "orteaf/internal/execution/allocator/policies/chunk_locator/direct_chunk_locator.h"
 #include "orteaf/internal/execution/allocator/policies/fast_free/fast_free_policies.h"
 #include "orteaf/internal/execution/allocator/policies/freelist/host_stack_freelist_policy.h"
